mocks: move object store to objectstore.go
